Qualification tool: Update speedup factor for few operators (#5490)
* Qualification tool: Update speedup factor for few operators


* addressed review comments

+++ tools/src/main/scala/com/nvidia/spark/rapids/tool/planparser/BatchScanExecParser.scala
@@ -37,7 +37,7 @@ case class BatchScanExecParser(
     // don't use the isExecSupported because we have finer grain.
     val score = ReadParser.calculateReadScoreRatio(readInfo, checker)
     val speedupFactor = checker.getSpeedupFactor(fullExecName)
-    val overallSpeedup = Math.max((speedupFactor * score).toInt, 1)
+    val overallSpeedup = Math.max((speedupFactor * score), 1.0)
 
     // TODO - add in parsing expressions - average speedup across?
     ExecInfo(sqlID, s"${node.name} ${readInfo.format}", "", overallSpeedup,
